intial commit

+++ update.py
@@ -1,9 +1,8 @@
 from logging import FileHandler, StreamHandler, INFO, basicConfig, error as log_error, info as log_info
-from os import path as ospath, environ, execl as osexecl
+from os import path as ospath, environ
 from subprocess import run as srun
 from requests import get as rget
 from dotenv import load_dotenv
-from sys import executable
 
 if ospath.exists('log.txt'):
     with open('log.txt', 'r+') as f:
@@ -37,27 +36,27 @@ try:
     if len(UPSTREAM_REPO) == 0:
        raise TypeError
 except:
-    UPSTREAM_REPO = "https://github.com/codewithweeb/WeebZone"
+    UPSTREAM_REPO = None
 try:
     if len(UPSTREAM_BRANCH) == 0:
        raise TypeError
 except:
-    UPSTREAM_BRANCH = 'update'
+    UPSTREAM_BRANCH = 'master'
 
-if ospath.exists('.git'):
-    srun(["rm", "-rf", ".git"])
+if UPSTREAM_REPO is not None:
+    if ospath.exists('.git'):
+        srun(["rm", "-rf", ".git"])
 
-update = srun([f"git init -q \
-                 && git config --global user.email codewithweeb@gmail.com \
-                 && git config --global user.name Karan \
-                 && git add . \
-                 && git commit -sm update -q \
-                 && git remote add origin {UPSTREAM_REPO} \
-                 && git fetch origin -q \
-                 && git reset --hard origin/{UPSTREAM_BRANCH} -q"], shell=True)
+    update = srun([f"git init -q \
+                     && git config --global user.email e.anastayyar@gmail.com \
+                     && git config --global user.name mltb \
+                     && git add . \
+                     && git commit -sm update -q \
+                     && git remote add origin {UPSTREAM_REPO} \
+                     && git fetch origin -q \
+                     && git reset --hard origin/{UPSTREAM_BRANCH} -q"], shell=True)
 
-if update.returncode == 0:
-    log_info('Successfully updated with latest commit from UPSTREAM_REPO')
-else:
-    log_error('Something went wrong while updating, check UPSTREAM_REPO if valid or not!')
-
+    if update.returncode == 0:
+        log_info('Successfully updated with latest commit from UPSTREAM_REPO')
+    else:
+        log_error('Something went wrong while updating, check UPSTREAM_REPO if valid or not!')
